+++ source/ShiftRegister/ShiftRegister.hpp
@@ -15,22 +15,22 @@ public:
   ~ShiftRegister() = default;
 
   auto step() -> int;
-  auto dataInput(uint64_t value) -> uint64_t;
-  [[nodiscard]] auto dataThrough() const -> uint64_t;
-  auto get(int index) -> uint64_t; // NOLINT 
+  auto dataInput(int value) -> int;
+  [[nodiscard]] auto dataThrough() const -> int;
+  auto get(int index) -> int; // NOLINT 
   [[nodiscard]] auto size() const -> int { return this->bits_; }
   auto activate() -> unsigned int;
   [[nodiscard]] auto getBias() const -> int;
   [[nodiscard]] auto calculateBias() const -> int;
-  auto currentInput() -> uint64_t;
+  auto currentInput() -> int;
   auto clearRegister(int registerID) -> int;
 
 private:
   int bits_ = DEFAULT_BIT_SIZE;
-  std::array<std::array<uint64_t, DEFAULT_BIT_SIZE>, REGISTER_COUNT> data_ = {};
+  std::array<std::array<int, DEFAULT_BIT_SIZE>, REGISTER_COUNT> data_ = {};
   unsigned int activeRegister_ = 0;
-  uint64_t currentInput_ = 0;
+  int currentInput_ = 0;
   int index_ = 0;
-  uint64_t currentThrough_ = 0;
+  int currentThrough_ = 0;
   int bias_ = 0;
 };
